project structure refactored, exceptions added
