+++ core/Quoridor.Core/Exceptions/PlayerLimitReachedException.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Quoridor.Core.Exceptions
+{
+    class PlayerLimitReachedException : ArgumentException
+    {
+        private const string MESSAGE = "Cannot add more players! Player limit reached: ";
+        public PlayerLimitReachedException(int limit) : base(MESSAGE + limit) { }
+    }
+}
