+++ core/Quoridor.Core/Models/Player.cs
@@ -1,4 +1,4 @@
-﻿namespace Quoridor.Core
+﻿namespace Quoridor.Core.Models
 {
     class Player
     {
