+++ core/Quoridor.Core/Models/Point.cs
@@ -1,4 +1,4 @@
-﻿namespace Quoridor.Core
+﻿namespace Quoridor.Core.Models
 {
     class Point
     {
